Updated and extended code to 'solve' Tic Tac Toe under the specified parameters
