+++ analysis/att1/trial_info.csv
@@ -1,0 +1,6 @@
+trial,Metric,Iters,Best_Iter,lr,x_tie_penalty,o_tie_penalty,gamma
+bb97dc4c,0.9999999999999944,117,100,0.00020071654256718154,0.8891175850956738,-0.14344294993766604,0.8681381250099209
+31b7349c,0.9986867463942116,117,112,1.1537276533002902e-05,0.37838556253374667,-0.20312599347084648,0.960783298104163
+6a00f222,0.9785739682438944,116,86,6.290241910993141e-05,0.4110560456465997,-0.42601942052495445,0.8228049023584615
+ac419e2f,0.946309487409066,115,46,0.00418574432892795,0.7489793532062917,-0.22283354722465032,0.9575636977794302
+9b197b43,0.6231954637491537,116,41,0.006953192454916885,0.3692113879819525,-0.9459224038321554,0.8880610654632924
