+++ analysis/replication/095eb04c.csv
@@ -1,0 +1,46 @@
+EnvSteps,Iters,TimeThisIter,TimeTotal,EpisodeLengthMean,EpisodeReturnMean,WinO,Tie,WinX,ReturnO,ReturnX
+4000.0,1,50.6695,50.6695,6.694,-0.1684,0.3153,0.1154,0.5692,-0.3722,0.2038
+8000.0,2,50.459,101.1285,6.2,-0.125,0.354,0.1012,0.5448,-0.2205,0.0955
+12000.0,3,50.233,151.3615,6.04,-0.1123,0.302,0.0852,0.6127,-0.3441,0.2319
+16000.0,4,49.8141,201.1756,5.868,-0.1046,0.3296,0.0756,0.5948,-0.3063,0.2017
+20000.0,5,49.2951,250.4707,5.802,-0.1199,0.2829,0.0893,0.6278,-0.37,0.25
+24000.0,6,48.7588,299.2295,5.6,-0.1072,0.3056,0.0878,0.6066,-0.3376,0.2304
+28000.0,7,48.4045,347.634,5.568,-0.1327,0.276,0.1089,0.6151,-0.3863,0.2537
+32000.0,8,47.4369,395.0709,5.698,-0.1148,0.3526,0.0705,0.577,-0.2914,0.1766
+36000.0,9,47.4278,442.4987,5.71,-0.1174,0.2523,0.102,0.6458,-0.4667,0.3493
+40000.0,10,47.7076,490.2063,5.85,-0.1608,0.2943,0.1245,0.5812,-0.3904,0.2296
+44000.0,11,47.9593,538.1657,5.844,-0.2041,0.2722,0.1511,0.5767,-0.3981,0.1939
+48000.0,12,48.2837,586.4494,5.9,-0.2041,0.2654,0.1462,0.5884,-0.4141,0.2099
+52000.0,13,48.4673,634.9167,6.134,-0.2756,0.2447,0.1958,0.5595,-0.4778,0.2022
+56000.0,14,48.7486,683.6654,6.264,-0.3649,0.1834,0.2591,0.5575,-0.5524,0.1876
+60000.0,15,48.7333,732.3987,6.288,-0.3776,0.1767,0.2747,0.5486,-0.5408,0.1632
+64000.0,16,48.0619,780.4606,6.286,-0.3623,0.1478,0.3004,0.5518,-0.5732,0.2108
+68000.0,17,48.5918,829.0524,6.5,-0.4389,0.1827,0.3419,0.4753,-0.4994,0.0606
+72000.0,18,49.0823,878.1347,6.732,-0.5231,0.167,0.4266,0.4064,-0.5155,-0.0075
+76000.0,19,47.4157,925.5504,6.884,-0.6277,0.114,0.4984,0.3876,-0.5739,-0.0539
+80000.0,20,45.9712,971.5216,6.902,-0.6047,0.1569,0.4819,0.3611,-0.5124,-0.0924
+84000.0,21,39.8837,1011.4053,7.056,-0.6787,0.1354,0.4929,0.3716,-0.5594,-0.1194
+88000.0,22,40.7341,1052.1394,6.972,-0.6354,0.1128,0.5157,0.3715,-0.5957,-0.0397
+92000.0,23,42.7345,1094.874,7.062,-0.6838,0.1624,0.5446,0.2929,-0.4979,-0.1859
+96000.0,24,42.1395,1137.0135,7.052,-0.6915,0.1278,0.5457,0.3265,-0.5237,-0.1677
+100000.0,25,43.8733,1180.8867,7.1,-0.7196,0.1037,0.557,0.3393,-0.5798,-0.1398
+104000.0,26,46.1889,1227.0756,7.174,-0.768,0.1296,0.5969,0.2735,-0.53,-0.238
+108000.0,27,47.8756,1274.9513,7.214,-0.7655,0.1306,0.5954,0.274,-0.5227,-0.2427
+112000.0,28,48.8364,1323.7876,7.268,-0.791,0.1071,0.6482,0.2447,-0.5355,-0.2555
+116000.0,29,48.6757,1372.4634,7.194,-0.7145,0.1407,0.5672,0.2921,-0.5292,-0.1852
+120000.0,30,48.6019,1421.0652,7.266,-0.7782,0.1204,0.605,0.2746,-0.5271,-0.2511
+124000.0,31,48.4994,1469.5646,7.234,-0.7782,0.1423,0.6001,0.2576,-0.5111,-0.2671
+128000.0,32,47.7823,1517.3469,7.344,-0.8625,0.1084,0.6863,0.2052,-0.5312,-0.3312
+132000.0,33,47.8582,1565.2051,7.19,-0.7221,0.1478,0.5795,0.2727,-0.4871,-0.2351
+136000.0,34,47.1795,1612.3846,7.362,-0.8165,0.1326,0.6363,0.2311,-0.5083,-0.3083
+140000.0,35,47.473,1659.8576,7.322,-0.8165,0.154,0.6633,0.1827,-0.4683,-0.3483
+144000.0,36,47.6395,1707.4971,7.328,-0.814,0.1436,0.6578,0.1986,-0.473,-0.341
+148000.0,37,47.9454,1755.4425,7.23,-0.7731,0.14,0.5977,0.2623,-0.4966,-0.2766
+152000.0,38,47.6364,1803.0789,7.248,-0.7961,0.1715,0.6357,0.1928,-0.4221,-0.3741
+156000.0,39,47.496,1850.5749,7.404,-0.8114,0.1476,0.627,0.2254,-0.4897,-0.3217
+160000.0,40,43.2419,1893.8167,7.39,-0.8522,0.1207,0.6813,0.198,-0.5141,-0.3381
+164000.0,41,39.5975,1933.4143,7.406,-0.8854,0.1272,0.6949,0.178,-0.4967,-0.3887
+168000.0,42,40.1037,1973.5179,7.386,-0.8599,0.1648,0.6704,0.1648,-0.4319,-0.4279
+172000.0,43,40.4926,2014.0105,7.452,-0.8752,0.1533,0.6868,0.16,-0.4716,-0.4036
+176000.0,44,43.3964,2057.4069,7.432,-0.9058,0.1095,0.7101,0.1805,-0.5109,-0.3949
+180000.0,45,44.828,2102.2349,7.436,-0.8727,0.1161,0.7118,0.1721,-0.4963,-0.3763
