+++ analysis/replication/2da66817.csv
@@ -1,0 +1,51 @@
+EnvSteps,Iters,TimeThisIter,TimeTotal,Tie,EpisodeLengthMean,EpisodeReturnMean,WinX,WinO,ReturnX,ReturnO
+4000.0,1,22.6812,22.6812,0.1406,6.63,0.0094,0.6065,0.2529,0.3327,-0.3233
+8000.0,2,34.0453,56.7265,0.0886,6.376,0.0065,0.5823,0.3291,0.2772,-0.2708
+12000.0,3,46.4464,103.1729,0.0894,6.142,0.0058,0.5697,0.3408,0.2189,-0.2131
+16000.0,4,46.9163,150.0892,0.0745,5.926,0.0053,0.5569,0.3686,0.2267,-0.2213
+20000.0,5,47.4739,197.5631,0.0916,5.626,0.0058,0.6445,0.2638,0.4029,-0.3971
+24000.0,6,48.7839,246.3469,0.0726,5.528,0.0056,0.6652,0.2622,0.3628,-0.3572
+28000.0,7,48.3491,294.696,0.1425,5.58,0.0084,0.6225,0.235,0.3842,-0.3758
+32000.0,8,48.7724,343.4684,0.1332,5.558,0.0088,0.5892,0.2776,0.3344,-0.3256
+36000.0,9,47.7302,391.1985,0.2405,5.926,0.0155,0.4819,0.2776,0.2077,-0.1923
+40000.0,10,47.2595,438.458,0.3789,6.342,0.0248,0.3972,0.2239,0.1864,-0.1616
+44000.0,11,47.0021,485.4601,0.5582,6.94,0.0359,0.2761,0.1657,0.108,-0.072
+48000.0,12,45.6953,531.1554,0.6498,7.174,0.0419,0.2128,0.1374,0.109,-0.067
+52000.0,13,45.7465,576.902,0.7493,7.378,0.0474,0.169,0.0817,0.1057,-0.0583
+56000.0,14,45.8262,622.7281,0.8395,7.592,0.0522,0.0919,0.0686,0.0201,0.0321
+60000.0,15,46.2351,668.9632,0.8559,7.684,0.055,0.1059,0.0381,0.0895,-0.0345
+64000.0,16,45.7462,714.7094,0.9015,7.794,0.0578,0.0463,0.0521,0.0349,0.0229
+68000.0,17,44.8576,759.567,0.9211,7.828,0.0587,0.0537,0.0252,0.0493,0.0093
+72000.0,18,44.1778,803.7448,0.9289,7.764,0.0584,0.0378,0.0333,0.0132,0.0452
+76000.0,19,44.4813,848.2261,0.9578,7.882,0.0606,0.0299,0.0123,0.0483,0.0123
+80000.0,20,45.6498,893.876,0.8718,7.806,0.0564,0.0311,0.0971,-0.0398,0.0962
+84000.0,21,46.6042,940.4801,0.9385,7.816,0.0587,0.0167,0.0448,0.0013,0.0573
+88000.0,22,47.0008,987.4809,0.9273,7.834,0.0597,0.0217,0.0509,0.0058,0.0538
+92000.0,23,40.5127,1027.9936,0.9535,7.896,0.0607,0.0135,0.0331,0.0103,0.0503
+96000.0,24,44.5293,1072.5229,0.9661,7.912,0.0615,0.0149,0.019,0.0307,0.0307
+100000.0,25,43.8205,1116.3434,0.9736,7.948,0.0617,0.0031,0.0233,0.0109,0.0509
+104000.0,26,43.9077,1160.2511,0.9563,7.912,0.0609,0.0038,0.0399,-0.0015,0.0625
+108000.0,27,43.6194,1203.8705,0.9547,7.898,0.0604,0.0142,0.0311,0.0102,0.0502
+112000.0,28,43.7674,1247.6379,0.9331,7.866,0.059,0.0227,0.0441,0.0075,0.0515
+116000.0,29,41.4602,1289.0981,0.9322,7.848,0.0588,0.0231,0.0447,-0.0006,0.0594
+120000.0,30,47.1286,1336.2268,0.9373,7.856,0.0598,0.0023,0.0605,-0.0241,0.0839
+124000.0,31,47.2535,1383.4803,0.9847,7.966,0.0627,0.0018,0.0135,0.0234,0.0394
+128000.0,32,46.9015,1430.3818,0.9803,7.956,0.0623,0.008,0.0117,0.0252,0.0372
+132000.0,33,46.2083,1476.5901,0.9611,7.906,0.0612,0.0055,0.0334,0.0066,0.0546
+136000.0,34,43.9136,1520.5037,0.9775,7.928,0.062,0.0056,0.0169,0.023,0.039
+140000.0,35,42.0031,1562.5068,0.9811,7.96,0.0625,0.0045,0.0144,0.0232,0.0392
+144000.0,36,42.6609,1605.1677,0.9561,7.888,0.0607,0.0016,0.0423,-0.0057,0.0663
+148000.0,37,43.6566,1648.8243,0.98,7.95,0.062,0.0005,0.0194,0.011,0.051
+152000.0,38,44.4985,1693.3228,0.9813,7.95,0.0622,0.0039,0.0148,0.0231,0.0391
+156000.0,39,45.0269,1738.3497,0.9776,7.932,0.062,0.0003,0.0221,0.007,0.055
+160000.0,40,45.5783,1783.928,0.9696,7.928,0.0618,0.0042,0.0263,0.0089,0.0529
+164000.0,41,45.6502,1829.5782,0.9505,7.898,0.0601,0.0041,0.0454,-0.016,0.076
+168000.0,42,45.6551,1875.2332,0.9682,7.93,0.0613,0.009,0.0228,0.0207,0.0407
+172000.0,43,45.5641,1920.7974,0.9817,7.954,0.0625,0.0067,0.0116,0.0272,0.0352
+176000.0,44,45.0856,1965.883,0.9831,7.95,0.0625,0.0011,0.0159,0.0192,0.0432
+180000.0,45,46.3129,2012.1959,0.9747,7.93,0.0617,0.0036,0.0216,0.0069,0.0549
+184000.0,46,45.4961,2057.692,0.9736,7.948,0.062,0.0003,0.0261,0.007,0.055
+188000.0,47,46.1192,2103.8113,0.9814,7.938,0.062,0.0011,0.0175,0.011,0.051
+192000.0,48,45.6974,2149.5087,0.9836,7.95,0.0623,0.0001,0.0164,0.0132,0.0492
+196000.0,49,45.0138,2194.5225,0.9881,7.968,0.0629,0.0006,0.0114,0.0254,0.0374
+200000.0,50,45.7323,2240.2548,0.9946,7.986,0.0631,0.0,0.0053,0.0256,0.0376
