+++ analysis/replication/8242f238.csv
@@ -1,0 +1,97 @@
+EnvSteps,Iters,TimeThisIter,TimeTotal,Tie,WinX,EpisodeLengthMean,EpisodeReturnMean,WinO,ReturnX,ReturnO
+4000.0,1,45.0405,45.0405,0.1181,0.6082,6.564,-0.0499,0.2736,0.263,-0.313
+8000.0,2,45.3791,90.4196,0.1105,0.5444,6.226,-0.0421,0.3451,0.1949,-0.2371
+12000.0,3,45.984,136.4035,0.1342,0.559,6.148,-0.0484,0.3068,0.2758,-0.3242
+16000.0,4,46.0798,182.4833,0.1555,0.5515,5.954,-0.0639,0.293,0.2,-0.264
+20000.0,5,45.9565,228.4399,0.153,0.5627,5.964,-0.0593,0.2843,0.2744,-0.3336
+24000.0,6,46.0661,274.5059,0.1497,0.5268,5.792,-0.0577,0.3235,0.1511,-0.2089
+28000.0,7,48.6743,323.1802,0.1481,0.5967,5.88,-0.0593,0.2551,0.3064,-0.3656
+32000.0,8,47.7729,370.9531,0.1921,0.539,5.868,-0.0717,0.2689,0.2121,-0.2839
+36000.0,9,46.5467,417.4998,0.207,0.5256,6.036,-0.0881,0.2674,0.2099,-0.2981
+40000.0,10,45.5029,463.0027,0.2782,0.509,6.318,-0.1162,0.2128,0.2159,-0.3321
+44000.0,11,43.7607,506.7634,0.3252,0.4483,6.406,-0.1256,0.2265,0.1312,-0.2568
+48000.0,12,43.61,550.3734,0.3562,0.4435,6.478,-0.1458,0.2002,0.1731,-0.3189
+52000.0,13,44.1127,594.4861,0.429,0.3922,6.564,-0.1669,0.1789,0.1446,-0.3114
+56000.0,14,44.8004,639.2865,0.433,0.3964,6.702,-0.1833,0.1706,0.1224,-0.3056
+60000.0,15,45.4053,684.6918,0.5526,0.2987,7.044,-0.223,0.1487,0.0525,-0.2755
+64000.0,16,45.5936,730.2853,0.6558,0.2423,7.244,-0.2574,0.1019,-0.0127,-0.2447
+68000.0,17,46.087,776.3723,0.6481,0.2305,7.242,-0.2527,0.1214,-0.0263,-0.2263
+72000.0,18,45.9231,822.2954,0.7054,0.1999,7.314,-0.2722,0.0946,-0.0221,-0.2501
+76000.0,19,46.0968,868.3923,0.7048,0.1795,7.41,-0.2815,0.1157,-0.0748,-0.2068
+80000.0,20,46.0037,914.3959,0.7613,0.1527,7.54,-0.3026,0.086,-0.0873,-0.2153
+84000.0,21,45.8081,960.204,0.7653,0.1573,7.52,-0.3026,0.0774,-0.0713,-0.2313
+88000.0,22,46.0989,1006.3029,0.7959,0.1433,7.518,-0.3104,0.0608,-0.0752,-0.2352
+92000.0,23,45.5565,1051.8594,0.8113,0.1221,7.556,-0.3166,0.0666,-0.1143,-0.2023
+96000.0,24,46.0882,1097.9476,0.771,0.1541,7.512,-0.3057,0.0749,-0.0889,-0.2169
+100000.0,25,45.9638,1143.9114,0.8318,0.1072,7.57,-0.3205,0.0609,-0.0983,-0.2223
+104000.0,26,45.9786,1189.89,0.807,0.132,7.612,-0.3174,0.061,-0.1007,-0.2167
+108000.0,27,46.0037,1235.8937,0.7919,0.1419,7.638,-0.3135,0.0662,-0.0688,-0.2448
+112000.0,28,41.5042,1277.3979,0.8294,0.0972,7.644,-0.3205,0.0734,-0.1343,-0.1863
+116000.0,29,40.9256,1318.3234,0.7793,0.1381,7.476,-0.3018,0.0827,-0.0849,-0.2169
+120000.0,30,40.6147,1358.9381,0.8037,0.1213,7.592,-0.3112,0.0751,-0.1056,-0.2056
+124000.0,31,44.5319,1403.47,0.7992,0.1016,7.588,-0.308,0.0992,-0.152,-0.156
+128000.0,32,47.0334,1450.5035,0.7523,0.1416,7.522,-0.3002,0.1062,-0.1001,-0.2001
+132000.0,33,48.4244,1498.9279,0.8025,0.1122,7.556,-0.3135,0.0853,-0.1288,-0.1848
+136000.0,34,48.5103,1547.4382,0.8091,0.1102,7.606,-0.3112,0.0807,-0.1096,-0.2016
+140000.0,35,48.8631,1596.3013,0.8212,0.113,7.602,-0.3151,0.0658,-0.1215,-0.1935
+144000.0,36,48.5757,1644.877,0.814,0.0955,7.64,-0.3174,0.0905,-0.1567,-0.1607
+148000.0,37,47.5019,1692.3789,0.7981,0.1332,7.56,-0.3096,0.0688,-0.1248,-0.1848
+152000.0,38,45.4085,1737.7874,0.8338,0.0822,7.63,-0.3244,0.084,-0.1742,-0.1502
+156000.0,39,43.1005,1780.888,0.8247,0.0835,7.64,-0.3197,0.0917,-0.1719,-0.1479
+160000.0,40,42.4269,1823.3149,0.7953,0.1218,7.53,-0.301,0.0829,-0.1265,-0.1745
+164000.0,41,42.8692,1866.1841,0.7959,0.099,7.584,-0.3112,0.1051,-0.1536,-0.1576
+168000.0,42,42.9401,1909.1241,0.8065,0.126,7.55,-0.3135,0.0675,-0.0848,-0.2288
+172000.0,43,43.1394,1952.2635,0.8714,0.0786,7.732,-0.3424,0.0499,-0.1692,-0.1732
+176000.0,44,43.6036,1995.8671,0.8668,0.0817,7.724,-0.3361,0.0516,-0.1501,-0.1861
+180000.0,45,44.3122,2040.1793,0.8508,0.0876,7.71,-0.3299,0.0616,-0.1349,-0.1949
+184000.0,46,45.9554,2086.1347,0.8635,0.0778,7.706,-0.3385,0.0586,-0.1452,-0.1932
+188000.0,47,46.6733,2132.808,0.8264,0.099,7.652,-0.3197,0.0747,-0.1479,-0.1719
+192000.0,48,46.7364,2179.5444,0.8337,0.094,7.568,-0.319,0.0722,-0.1375,-0.1815
+196000.0,49,46.8342,2226.3786,0.8021,0.101,7.612,-0.3112,0.0969,-0.1376,-0.1736
+200000.0,50,44.5481,2270.9266,0.8223,0.0791,7.642,-0.3229,0.0986,-0.1694,-0.1534
+204000.0,51,49.0683,2319.995,0.7992,0.1295,7.578,-0.3088,0.0713,-0.1024,-0.2064
+208000.0,52,48.4676,2368.4626,0.7788,0.1462,7.532,-0.3049,0.075,-0.0785,-0.2265
+212000.0,53,47.0101,2415.4727,0.7908,0.1245,7.514,-0.3073,0.0848,-0.1216,-0.1856
+216000.0,54,43.0274,2458.5001,0.7732,0.1126,7.532,-0.3049,0.1142,-0.1665,-0.1385
+220000.0,55,41.6983,2500.1984,0.7999,0.1362,7.53,-0.3049,0.0639,-0.0745,-0.2305
+224000.0,56,42.9915,2543.1899,0.7389,0.1484,7.552,-0.2909,0.1127,-0.0994,-0.1914
+228000.0,57,43.539,2586.7289,0.773,0.1046,7.506,-0.3002,0.1223,-0.1481,-0.1521
+232000.0,58,44.251,2630.9799,0.7407,0.1573,7.356,-0.2839,0.102,-0.0619,-0.2219
+236000.0,59,43.9249,2674.9047,0.7418,0.1703,7.492,-0.2909,0.0879,-0.0594,-0.2314
+240000.0,60,45.2461,2720.1509,0.7861,0.1324,7.452,-0.2979,0.0816,-0.073,-0.225
+244000.0,61,45.5863,2765.7372,0.7912,0.1368,7.52,-0.3073,0.072,-0.0856,-0.2216
+248000.0,62,46.4433,2812.1804,0.7924,0.1577,7.494,-0.3018,0.0499,-0.0249,-0.2769
+252000.0,63,47.0137,2859.1941,0.7328,0.1957,7.454,-0.2917,0.0715,-0.0578,-0.2338
+256000.0,64,46.7848,2905.9789,0.8007,0.1336,7.616,-0.3174,0.0657,-0.1007,-0.2167
+260000.0,65,47.2927,2953.2716,0.7806,0.134,7.446,-0.3026,0.0854,-0.1073,-0.1953
+264000.0,66,46.3048,2999.5765,0.8164,0.1176,7.606,-0.319,0.0659,-0.1175,-0.2015
+268000.0,67,43.4504,3043.0268,0.8028,0.1102,7.516,-0.3065,0.087,-0.1072,-0.1992
+272000.0,68,39.1686,3082.1954,0.7998,0.146,7.534,-0.3135,0.0542,-0.0768,-0.2368
+276000.0,69,38.3155,3120.5109,0.8436,0.091,7.664,-0.3307,0.0654,-0.1333,-0.1973
+280000.0,70,38.4666,3158.9774,0.841,0.1096,7.586,-0.319,0.0495,-0.1095,-0.2095
+284000.0,71,45.6539,3204.6313,0.7744,0.1577,7.5,-0.2971,0.0679,-0.0546,-0.2426
+288000.0,72,47.9101,3252.5415,0.8233,0.1112,7.548,-0.3143,0.0655,-0.1151,-0.1991
+292000.0,73,48.5342,3301.0757,0.7783,0.1268,7.51,-0.301,0.0949,-0.0945,-0.2065
+296000.0,74,48.083,3349.1587,0.7365,0.1727,7.454,-0.2909,0.0908,-0.0714,-0.2194
+300000.0,75,48.2523,3397.411,0.7374,0.1655,7.414,-0.2909,0.0971,-0.0794,-0.2114
+304000.0,76,48.1123,3445.5234,0.8282,0.0981,7.562,-0.3174,0.0736,-0.1527,-0.1647
+308000.0,77,48.6436,3494.167,0.7849,0.1441,7.564,-0.308,0.071,-0.096,-0.212
+312000.0,78,48.6751,3542.8421,0.7966,0.115,7.498,-0.3073,0.0883,-0.1536,-0.1536
+316000.0,79,49.2761,3592.1183,0.7949,0.116,7.472,-0.3026,0.0891,-0.1193,-0.1833
+320000.0,80,49.3562,3641.4744,0.7966,0.1107,7.578,-0.3127,0.0927,-0.1304,-0.1824
+324000.0,81,48.5475,3690.022,0.8313,0.0933,7.592,-0.3166,0.0754,-0.1463,-0.1703
+328000.0,82,47.6901,3737.7121,0.8114,0.0952,7.584,-0.3135,0.0934,-0.1448,-0.1688
+332000.0,83,47.0202,3784.7323,0.8078,0.1196,7.616,-0.3213,0.0726,-0.1207,-0.2007
+336000.0,84,46.069,3830.8013,0.8399,0.1214,7.674,-0.3268,0.0387,-0.0934,-0.2334
+340000.0,85,45.3237,3876.125,0.8342,0.0818,7.674,-0.3268,0.084,-0.1574,-0.1694
+344000.0,86,43.833,3919.9581,0.851,0.082,7.692,-0.3275,0.067,-0.1478,-0.1798
+348000.0,87,42.7521,3962.7101,0.8293,0.0803,7.684,-0.3283,0.0905,-0.1742,-0.1542
+352000.0,88,42.9024,4005.6125,0.8029,0.1249,7.562,-0.3104,0.0722,-0.0752,-0.2352
+356000.0,89,41.5792,4047.1916,0.8052,0.141,7.542,-0.3104,0.0538,-0.0752,-0.2352
+360000.0,90,40.45,4087.6416,0.8052,0.1117,7.532,-0.3143,0.0831,-0.1391,-0.1751
+364000.0,91,43.2905,4130.9321,0.8011,0.113,7.568,-0.3143,0.0859,-0.0991,-0.2151
+368000.0,92,47.4963,4178.4285,0.7757,0.159,7.55,-0.3127,0.0653,-0.0584,-0.2544
+372000.0,93,47.8313,4226.2597,0.8089,0.1187,7.59,-0.3158,0.0723,-0.1319,-0.1839
+376000.0,94,47.5084,4273.7681,0.715,0.1681,7.446,-0.2901,0.1169,-0.1091,-0.1811
+380000.0,95,47.7117,4321.4798,0.7855,0.1086,7.538,-0.3073,0.1059,-0.1376,-0.1696
+384000.0,96,47.6758,4369.1556,0.8276,0.0981,7.664,-0.3229,0.0743,-0.1334,-0.1894
