+++ analysis/replication/98cc34f1.csv
@@ -1,0 +1,89 @@
+EnvSteps,Iters,TimeThisIter,TimeTotal,WinX,EpisodeReturnMean,Tie,WinO,EpisodeLengthMean,ReturnO,ReturnX
+4000.0,1,49.1776,49.1776,0.6,-0.1765,0.1193,0.2807,6.71,-0.3823,0.2057
+8000.0,2,48.6051,97.7827,0.6535,-0.1149,0.0704,0.2761,6.208,-0.3834,0.2686
+12000.0,3,48.6711,146.4538,0.5483,-0.1149,0.0741,0.3777,5.98,-0.2474,0.1326
+16000.0,4,48.7995,195.2533,0.6346,-0.1093,0.0649,0.3004,5.822,-0.3886,0.2794
+20000.0,5,48.8481,244.1014,0.5995,-0.1177,0.0721,0.3284,5.774,-0.3308,0.2132
+24000.0,6,50.5184,294.6198,0.638,-0.07,0.0536,0.3083,5.602,-0.365,0.295
+28000.0,7,49.5815,344.2013,0.5992,-0.0757,0.0545,0.3463,5.676,-0.3118,0.2362
+32000.0,8,48.3928,392.5941,0.6333,-0.1009,0.063,0.3037,5.678,-0.3824,0.2816
+36000.0,9,48.0859,440.68,0.562,-0.1233,0.0857,0.3523,5.696,-0.3016,0.1784
+40000.0,10,46.8811,487.5611,0.5386,-0.227,0.1611,0.3003,5.898,-0.3635,0.1365
+44000.0,11,46.5283,534.0894,0.5282,-0.1989,0.1468,0.325,5.914,-0.3135,0.1145
+48000.0,12,45.863,579.9524,0.4753,-0.2886,0.1963,0.3284,6.08,-0.3143,0.0257
+52000.0,13,45.5508,625.5031,0.4517,-0.311,0.2243,0.3239,6.178,-0.2975,-0.0135
+56000.0,14,45.0476,670.5507,0.425,-0.4259,0.2878,0.2872,6.404,-0.3729,-0.0529
+60000.0,15,45.1096,715.6603,0.4346,-0.4539,0.2936,0.2718,6.49,-0.363,-0.091
+64000.0,16,44.9217,760.582,0.4015,-0.538,0.3531,0.2454,6.644,-0.405,-0.133
+68000.0,17,44.7808,805.3628,0.37,-0.5884,0.4048,0.2252,6.83,-0.4142,-0.1742
+72000.0,18,45.5105,850.8733,0.3027,-0.6556,0.5105,0.1868,6.86,-0.4678,-0.1878
+76000.0,19,44.5466,895.42,0.3275,-0.6781,0.4828,0.1898,7.022,-0.499,-0.179
+80000.0,20,44.9857,940.4057,0.3095,-0.7089,0.5046,0.1858,7.038,-0.4844,-0.2244
+84000.0,21,39.4754,979.8811,0.3046,-0.7873,0.5711,0.1244,7.208,-0.5437,-0.2437
+88000.0,22,45.2483,1025.1294,0.2748,-0.7957,0.5555,0.1697,7.156,-0.5019,-0.2939
+92000.0,23,46.7349,1071.8643,0.2648,-0.7873,0.5827,0.1525,7.156,-0.5277,-0.2597
+96000.0,24,47.5623,1119.4265,0.2588,-0.8322,0.5965,0.1447,7.198,-0.5541,-0.2781
+100000.0,25,48.5687,1167.9952,0.2176,-0.8602,0.621,0.1613,7.3,-0.5121,-0.3481
+104000.0,26,49.7776,1217.7728,0.2176,-0.8854,0.6206,0.1618,7.32,-0.5147,-0.3707
+108000.0,27,48.436,1266.2088,0.2625,-0.821,0.5961,0.1414,7.252,-0.5445,-0.2765
+112000.0,28,48.2868,1314.4955,0.2426,-0.8966,0.6364,0.121,7.33,-0.5643,-0.3323
+116000.0,29,47.8417,1362.3372,0.2549,-0.8238,0.5892,0.1558,7.22,-0.5119,-0.3119
+120000.0,30,45.222,1407.5592,0.2617,-0.8238,0.5911,0.1472,7.142,-0.4919,-0.3319
+124000.0,31,41.467,1449.0263,0.2634,-0.8294,0.5884,0.1482,7.192,-0.4947,-0.3347
+128000.0,32,42.0083,1491.0346,0.2276,-0.8714,0.6144,0.158,7.29,-0.4977,-0.3737
+132000.0,33,42.4231,1533.4577,0.2398,-0.8994,0.6542,0.1061,7.318,-0.5557,-0.3437
+136000.0,34,43.7542,1577.2119,0.2223,-0.9078,0.6507,0.1271,7.324,-0.5499,-0.3579
+140000.0,35,43.3979,1620.6098,0.1659,-0.9246,0.6632,0.1709,7.264,-0.4183,-0.5063
+144000.0,36,44.4522,1665.062,0.197,-0.8602,0.6251,0.1779,7.238,-0.4521,-0.4081
+148000.0,37,46.0872,1711.1492,0.2239,-0.9078,0.645,0.1311,7.266,-0.5539,-0.3539
+152000.0,38,46.5428,1757.692,0.1543,-0.9498,0.7037,0.142,7.348,-0.5009,-0.4489
+156000.0,39,47.3211,1805.0132,0.1539,-0.933,0.6792,0.1669,7.3,-0.4485,-0.4845
+160000.0,40,47.157,1852.1701,0.2161,-0.8686,0.6279,0.156,7.228,-0.5103,-0.3583
+164000.0,41,47.2177,1899.3879,0.2298,-0.891,0.6501,0.1201,7.162,-0.5335,-0.3575
+168000.0,42,47.1038,1946.4917,0.2006,-0.8854,0.6456,0.1538,7.28,-0.4667,-0.4187
+172000.0,43,44.8549,1991.3466,0.1764,-0.8938,0.6456,0.178,7.198,-0.4529,-0.4409
+176000.0,44,49.1116,2040.4583,0.2158,-0.849,0.6065,0.1777,7.122,-0.4705,-0.3785
+180000.0,45,49.1025,2089.5607,0.1798,-0.9302,0.6499,0.1703,7.29,-0.4611,-0.4691
+184000.0,46,48.293,2137.8537,0.1899,-0.8854,0.6121,0.1979,7.23,-0.4387,-0.4467
+188000.0,47,46.9245,2184.7783,0.1748,-0.8854,0.647,0.1782,7.262,-0.4307,-0.4547
+192000.0,48,44.2164,2228.9947,0.1618,-0.9975,0.7268,0.1114,7.402,-0.5507,-0.4467
+196000.0,49,43.2858,2272.2805,0.1763,-0.9106,0.6577,0.166,7.33,-0.4813,-0.4293
+200000.0,50,43.7284,2316.0089,0.1929,-0.9274,0.654,0.1531,7.246,-0.5017,-0.4257
+204000.0,51,44.2267,2360.2356,0.1961,-0.891,0.6276,0.1763,7.27,-0.4455,-0.4455
+208000.0,52,44.4257,2404.6614,0.1621,-0.8686,0.6455,0.1924,7.218,-0.4423,-0.4263
+212000.0,53,44.1255,2448.7869,0.2419,-0.8434,0.5913,0.1668,7.202,-0.4637,-0.3797
+216000.0,54,44.5783,2493.3652,0.2361,-0.7845,0.5659,0.198,7.12,-0.4323,-0.3523
+220000.0,55,44.9952,2538.3604,0.2213,-0.835,0.5971,0.1816,7.234,-0.4655,-0.3695
+224000.0,56,46.0532,2584.4136,0.2407,-0.7901,0.5658,0.1936,7.1,-0.4311,-0.3591
+228000.0,57,46.5646,2630.9782,0.2023,-0.8266,0.5906,0.2071,7.222,-0.4113,-0.4153
+232000.0,58,47.3012,2678.2794,0.2117,-0.7621,0.5541,0.2342,7.084,-0.3731,-0.3891
+236000.0,59,47.617,2725.8964,0.2863,-0.7677,0.5401,0.1736,7.008,-0.4519,-0.3159
+240000.0,60,47.9491,2773.8455,0.2265,-0.8097,0.572,0.2016,7.132,-0.4349,-0.3749
+244000.0,61,48.4453,2822.2907,0.2115,-0.8462,0.6049,0.1836,7.176,-0.4751,-0.3711
+248000.0,62,48.6703,2870.961,0.2543,-0.8153,0.5914,0.1543,7.144,-0.5137,-0.3017
+252000.0,63,48.5006,2919.4616,0.234,-0.8294,0.6191,0.1469,7.202,-0.5067,-0.3227
+256000.0,64,48.2987,2967.7603,0.2389,-0.891,0.6296,0.1315,7.216,-0.5455,-0.3455
+260000.0,65,46.7394,3014.4998,0.2229,-0.891,0.622,0.1551,7.23,-0.5015,-0.3895
+264000.0,66,44.8499,3059.3497,0.2408,-0.8826,0.6198,0.1394,7.21,-0.5353,-0.3473
+268000.0,67,41.864,3101.2137,0.2491,-0.8434,0.6173,0.1337,7.172,-0.5317,-0.3117
+272000.0,68,42.3245,3143.5382,0.2765,-0.7369,0.526,0.1975,6.95,-0.4224,-0.3144
+276000.0,69,40.7989,3184.3371,0.2784,-0.7285,0.5206,0.201,6.956,-0.4522,-0.2762
+280000.0,70,39.4987,3223.8357,0.3253,-0.6753,0.4802,0.1945,6.88,-0.4556,-0.2196
+284000.0,71,44.6216,3268.4574,0.3496,-0.6332,0.4385,0.2119,6.874,-0.4206,-0.2126
+288000.0,72,47.7615,3316.2188,0.2927,-0.6781,0.4915,0.2158,6.942,-0.419,-0.259
+292000.0,73,48.482,3364.7008,0.3571,-0.6809,0.4835,0.1595,6.852,-0.5584,-0.1224
+296000.0,74,48.573,3413.2738,0.383,-0.6332,0.4543,0.1626,6.76,-0.5286,-0.1046
+300000.0,75,48.4586,3461.7324,0.3569,-0.6865,0.5017,0.1413,6.768,-0.5892,-0.0972
+304000.0,76,48.1155,3509.8479,0.3222,-0.7369,0.5439,0.1339,6.904,-0.5384,-0.1984
+308000.0,77,47.7553,3557.6031,0.359,-0.7285,0.4952,0.1457,6.928,-0.5882,-0.1402
+312000.0,78,47.2388,3604.8419,0.3134,-0.7789,0.5694,0.1172,7.044,-0.5935,-0.1855
+316000.0,79,47.467,3652.3089,0.289,-0.7621,0.5498,0.1612,6.982,-0.5131,-0.2491
+320000.0,80,47.8424,3700.1512,0.2946,-0.8041,0.5675,0.1379,7.1,-0.5961,-0.2081
+324000.0,81,47.7647,3747.916,0.302,-0.8434,0.6054,0.0926,7.106,-0.6237,-0.2197
+328000.0,82,48.1715,3796.0875,0.3067,-0.7453,0.5445,0.1488,6.914,-0.5327,-0.2127
+332000.0,83,48.7433,3844.8307,0.3051,-0.8182,0.5887,0.1063,7.072,-0.6091,-0.2091
+336000.0,84,48.0792,3892.9099,0.3273,-0.8406,0.5792,0.0934,7.156,-0.6203,-0.2203
+340000.0,85,47.0254,3939.9353,0.3318,-0.8041,0.5489,0.1192,7.09,-0.5841,-0.2201
+344000.0,86,46.5811,3986.5164,0.3065,-0.7677,0.5382,0.1553,6.978,-0.5119,-0.2559
+348000.0,87,44.8642,4031.3806,0.3517,-0.8125,0.5392,0.1091,7.096,-0.6343,-0.1783
+352000.0,88,43.2654,4074.646,0.2906,-0.7929,0.5927,0.1167,7.036,-0.5665,-0.2265
