+++ analysis/replication/a294744a.csv
@@ -1,0 +1,28 @@
+EnvSteps,Iters,TimeThisIter,TimeTotal,EpisodeReturnMean,WinO,Tie,EpisodeLengthMean,WinX,ReturnO,ReturnX
+4000.0,1,44.8822,44.8822,0.0724,0.3039,0.1,6.572,0.5961,-0.2238,0.2962
+8000.0,2,45.8997,90.7819,0.0644,0.2982,0.099,6.294,0.6028,-0.2678,0.3322
+12000.0,3,45.7498,136.5316,0.0657,0.3392,0.098,6.31,0.5628,-0.2332,0.2988
+16000.0,4,45.9893,182.521,0.0402,0.31,0.0695,5.866,0.6205,-0.3079,0.3481
+20000.0,5,46.0297,228.5506,0.0483,0.3332,0.0677,5.798,0.5991,-0.2519,0.3001
+24000.0,6,46.1101,274.6608,0.0523,0.2649,0.0949,5.61,0.6402,-0.3479,0.4001
+28000.0,7,46.0599,320.7207,0.0577,0.3287,0.0894,5.59,0.5819,-0.2732,0.3308
+32000.0,8,46.7343,367.4549,0.0925,0.2635,0.1327,5.694,0.6038,-0.2997,0.3923
+36000.0,9,46.9539,414.4088,0.1234,0.2333,0.1888,5.738,0.5779,-0.2943,0.4177
+40000.0,10,47.516,461.9249,0.185,0.2739,0.2655,5.972,0.4607,-0.1195,0.3045
+44000.0,11,45.9067,507.8316,0.2427,0.3146,0.3613,6.492,0.3241,0.1153,0.1273
+48000.0,12,45.092,552.9236,0.3352,0.2476,0.5041,6.82,0.2483,0.1636,0.1716
+52000.0,13,43.5885,596.5121,0.4492,0.1537,0.664,7.102,0.1823,0.2026,0.2466
+56000.0,14,43.2312,639.7433,0.5323,0.0959,0.7796,7.49,0.1246,0.2321,0.3001
+60000.0,15,43.861,683.6042,0.5497,0.0605,0.8232,7.554,0.1163,0.2189,0.3309
+64000.0,16,44.6436,728.2478,0.6047,0.0347,0.896,7.768,0.0694,0.2763,0.3283
+68000.0,17,44.7029,772.9508,0.6154,0.0265,0.9055,7.774,0.068,0.2777,0.3377
+72000.0,18,45.072,818.0228,0.6329,0.0171,0.9518,7.854,0.0311,0.3044,0.3284
+76000.0,19,45.6432,863.666,0.6449,0.0147,0.9557,7.896,0.0296,0.3085,0.3365
+80000.0,20,44.3823,908.0483,0.6409,0.032,0.9446,7.878,0.0235,0.3365,0.3045
+84000.0,21,43.972,952.0203,0.653,0.0141,0.9799,7.928,0.006,0.3325,0.3205
+88000.0,22,45.6104,997.6307,0.6476,0.0172,0.9679,7.916,0.0148,0.3298,0.3178
+92000.0,23,45.382,1043.0127,0.6516,0.0153,0.9737,7.916,0.011,0.3298,0.3218
+96000.0,24,45.7585,1088.7713,0.657,0.0054,0.9788,7.956,0.0158,0.3165,0.3405
+100000.0,25,45.646,1134.4173,0.6543,0.0116,0.9776,7.938,0.0108,0.3232,0.3312
+104000.0,26,45.6855,1180.1028,0.6624,0.006,0.9886,7.978,0.0054,0.3312,0.3312
+108000.0,27,45.4425,1225.5453,0.665,0.0005,0.997,7.976,0.0025,0.3245,0.3405
