+++ analysis/replication/af610087.csv
@@ -1,0 +1,22 @@
+EnvSteps,Iters,TimeThisIter,TimeTotal,EpisodeLengthMean,WinX,WinO,EpisodeReturnMean,Tie,ReturnX,ReturnO
+4000.0,1,48.9096,48.9096,6.704,0.614,0.2606,0.1004,0.1255,0.3662,-0.2658
+8000.0,2,49.1149,98.0245,6.502,0.5361,0.2985,0.1047,0.1654,0.3023,-0.1977
+12000.0,3,49.3459,147.3704,6.184,0.5653,0.3114,0.0932,0.1233,0.3166,-0.2234
+16000.0,4,50.0966,197.467,6.074,0.5794,0.2845,0.1004,0.1361,0.3542,-0.2538
+20000.0,5,50.8644,248.3314,5.93,0.586,0.2532,0.1233,0.1607,0.3897,-0.2663
+24000.0,6,50.0741,298.4055,5.884,0.5811,0.2273,0.1391,0.1915,0.4195,-0.2805
+28000.0,7,50.3376,348.743,6.204,0.5176,0.2049,0.2007,0.2776,0.4284,-0.2276
+32000.0,8,50.0137,398.7567,6.14,0.518,0.1879,0.2136,0.2941,0.4568,-0.2432
+36000.0,9,49.3216,448.0783,6.45,0.4096,0.2076,0.2825,0.3828,0.3352,-0.0528
+40000.0,10,48.4595,496.5379,6.808,0.3005,0.1912,0.3742,0.5083,0.3211,0.0531
+44000.0,11,48.4753,545.0132,7.056,0.2254,0.1372,0.456,0.6375,0.312,0.144
+48000.0,12,48.7452,593.7584,7.398,0.1247,0.1132,0.542,0.7622,0.279,0.263
+52000.0,13,47.5506,641.309,7.65,0.1003,0.0562,0.6165,0.8435,0.3403,0.2763
+56000.0,14,47.1653,688.4743,7.73,0.0747,0.0429,0.6381,0.8824,0.337,0.301
+60000.0,15,47.2696,735.7439,7.892,0.021,0.0186,0.6868,0.9604,0.3374,0.3494
+64000.0,16,46.8102,782.5541,7.802,0.0316,0.0442,0.6567,0.9242,0.3203,0.3363
+68000.0,17,46.6266,829.1806,7.944,0.0166,0.0121,0.6997,0.9713,0.3579,0.3419
+72000.0,18,45.4697,874.6503,7.918,0.0163,0.017,0.6925,0.9667,0.3443,0.3483
+76000.0,19,44.0207,918.671,7.964,0.0119,0.0035,0.7069,0.9846,0.3594,0.3474
+80000.0,20,43.6847,962.3557,7.956,0.0073,0.0068,0.7026,0.9858,0.3473,0.3553
+84000.0,21,44.0233,1006.379,7.976,0.0034,0.0048,0.7097,0.9917,0.3569,0.3529
