+++ analysis/replication/b150e5c8.csv
@@ -1,0 +1,22 @@
+EnvSteps,Iters,TimeThisIter,TimeTotal,EpisodeLengthMean,Tie,WinX,WinO,EpisodeReturnMean,ReturnX,ReturnO
+4000.0,1,36.5182,36.5182,6.634,0.1381,0.5908,0.271,0.1639,0.432,-0.268
+8000.0,2,45.6571,82.1754,6.462,0.1643,0.56,0.2757,0.2334,0.4007,-0.1673
+12000.0,3,46.1885,128.3639,6.31,0.1943,0.5456,0.2602,0.2834,0.3897,-0.1063
+16000.0,4,46.9792,175.343,6.406,0.2579,0.486,0.256,0.3946,0.4053,-0.0107
+20000.0,5,47.0609,222.4039,6.424,0.3527,0.4415,0.2058,0.4751,0.4876,-0.0124
+24000.0,6,47.7417,270.1456,6.632,0.4849,0.3717,0.1434,0.6502,0.5451,0.1051
+28000.0,7,47.6982,317.8438,6.856,0.5566,0.3519,0.0915,0.7752,0.6376,0.1376
+32000.0,8,47.2052,365.049,7.09,0.6225,0.2876,0.0899,0.8864,0.6172,0.2692
+36000.0,9,46.993,412.042,7.21,0.6876,0.2521,0.0603,0.9559,0.6499,0.3059
+40000.0,10,46.0053,458.0472,7.33,0.7365,0.2139,0.0496,1.0364,0.6642,0.3722
+44000.0,11,46.0053,504.0525,7.54,0.8007,0.1563,0.043,1.1142,0.6831,0.4311
+48000.0,12,45.8516,549.9042,7.604,0.8376,0.1383,0.024,1.1531,0.6986,0.4546
+52000.0,13,46.2813,596.1855,7.708,0.8927,0.0955,0.0118,1.2282,0.6981,0.5301
+56000.0,14,46.0719,642.2574,7.778,0.9109,0.0787,0.0105,1.2671,0.7015,0.5655
+60000.0,15,45.8884,688.1457,7.846,0.9254,0.0615,0.0131,1.3004,0.6942,0.6062
+64000.0,16,45.4113,733.5571,7.854,0.9369,0.0593,0.0038,1.3032,0.7016,0.6016
+68000.0,17,45.4015,778.9585,7.908,0.9676,0.0302,0.0023,1.3365,0.6983,0.6383
+72000.0,18,45.2822,824.2408,7.942,0.9765,0.0217,0.0019,1.356,0.69,0.666
+76000.0,19,44.9962,869.237,7.972,0.9861,0.0092,0.0047,1.3754,0.6897,0.6857
+80000.0,20,45.1652,914.4022,7.972,0.9854,0.0143,0.0004,1.3754,0.6977,0.6777
+84000.0,21,46.6742,961.0764,7.996,0.9972,0.0028,0.0,1.3865,0.6953,0.6913
