+++ analysis/replication/d0b80b71.png
