+++ analysis/replication/trial_info.csv
@@ -1,0 +1,11 @@
+trial,Metric,Iters,Best_Iter,lr,tie_penalty,gamma
+b150e5c8,0.9971740993314576,21,21,0.00028680243649277786,0.694662071094744,0.984343331592276
+a294744a,0.9969647080737128,27,27,6.563820049480557e-05,0.3351991760505908,0.9198887413305485
+2da66817,0.994607515190224,50,50,9.16829806363971e-05,0.03174324356754088,0.916639034876105
+af610087,0.9917480374002904,21,21,0.00012249873224163968,0.3584575210703531,0.954354300759849
+ca0eceae,0.957779324053022,117,102,3.234996176722559e-05,-0.7172583647875148,0.9539707008678788
+2dd2537e,0.9086573735549264,116,97,0.0020969750984383543,-0.12362702207911913,0.9356379087689773
+8242f238,0.871434971913041,96,43,0.0016641559256606873,-0.19496502955576323,0.9551933705407369
+98cc34f1,0.7267810556994212,88,48,0.002364565331713159,-0.7004709335628887,0.9476879582618583
+095eb04c,0.7118417033109606,45,45,0.003476168392400587,-0.6379072401826693,0.840285356363364
+9004554f,0.1280261124229172,1,1,4.7973917731170575e-05,-0.954755269690678,0.9777311400807173
