--- articles/00FF0000/analysis/trial_info.csv
@@ -1,6 +1,0 @@
-trial,Metric,Iters,Best_Iter,lr,tie_penalty,gamma
-d0b80b71,0.973551274294054,241,236,0.00010720296175728038,-0.23893360580501288,0.9225237757047842
-4f927e2f,0.9551402959799988,242,236,0.0004450688969770199,0.6819230722540293,0.823800570678455
-1bdac927,0.9382096332414316,243,243,0.00025551151358014167,-0.47623898294514233,0.9685848894178615
-2d565d0a,0.8453946154525182,241,227,0.0010058697342042215,-0.2669636577264056,0.9662118843025685
-5dc51969,0.8108531799542288,243,243,0.00039223516355616145,-0.717031638240065,0.9690992763742861
