+++ requirements.txt
@@ -1,0 +1,19 @@
+ray[rllib]
+gymnasium==1.0.0
+pettingzoo
+networkx==3.2.1
+hyperopt
+onnx
+onnxruntime
+pandas==2.2.3
+torch
+numpy==1.26.4
+pandas==2.2.3
+argparse
+
+dm_tree 
+fsspec 
+pyarrow==14.0.2 
+scipy 
+lz4 
+aiohttp_cors
