+++ training.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import os
+import argparse
 from pathlib import Path
 from typing import TYPE_CHECKING
 
@@ -8,6 +9,7 @@ import ray
 
 from ray import tune, train
 from ray.rllib.algorithms import PPOConfig
+from ray.rllib.core.rl_module.default_model_config import DefaultModelConfig
 from ray.rllib.core.rl_module.rl_module import RLModuleSpec
 from ray.rllib.core.rl_module.multi_rl_module import MultiRLModuleSpec
 from ray.rllib.env import ParallelPettingZooEnv
@@ -19,7 +21,10 @@ from ray.tune.search import ConcurrencyLimiter
 from custom_metric import Outcomes
 from enviornment import MultiAgentTicTacToe
 from shared.ray.progress_report import CustomReporter
-from shared.ray.action_masking import ActionMaskingTorchRLModule
+from ray.rllib.examples.rl_modules.classes.action_masking_rlm import (
+    ActionMaskingTorchRLModule,
+)
+#from shared.ray.action_masking import ActionMaskingTorchRLModule
 
 if TYPE_CHECKING:
     from ray.tune.experiment import Trial
@@ -42,7 +47,7 @@ def trail_dirname_creator(trial: Trial) -> str:
     return f'{trial.trial_id}'
 
 
-def main():
+def main(args):
     """
     Train agent X and agent O against one another. X will be forced
     to play a random first move, which helps O experience the entire
@@ -54,7 +59,7 @@ def main():
     :return:
     """
 
-    experiment_name = 'Random-First-Move'
+    experiment_name = f'Settings: {int(100*args.x_win_reward)} {int(100*args.o_win_reward)} {int(100*args.x_lose_reward)} {int(100*args.o_lose_reward)} {int(100*args.stop_tie_threshold)}'
     max_iter_individual = 300
     max_time_total = 60 * 90
     grace_period_iter = 15
@@ -70,10 +75,10 @@ def main():
     register_env('tic-tac-toe', env_creator)
     config = (
         PPOConfig()
-        .api_stack(
-            enable_env_runner_and_connector_v2=True,
-            enable_rl_module_and_learner=True
-        )
+        #.api_stack(
+        #    enable_env_runner_and_connector_v2=True,
+        #    enable_rl_module_and_learner=True
+        #)
         .reporting(
             metrics_num_episodes_for_smoothing=1000
         )
@@ -83,8 +88,13 @@ def main():
         .environment(
             env='tic-tac-toe',
             env_config={
-                'tie_penalty': tune.uniform(-1, 1),
+                'x_tie_penalty': tune.uniform(args.x_lose_reward, args.x_win_reward),
+                'o_tie_penalty': tune.uniform(args.o_lose_reward, args.o_win_reward),
                 'random_first': True,
+                'x_win_reward': args.x_win_reward,
+                'x_lose_reward': args.x_lose_reward,
+                'o_win_reward': args.o_win_reward,
+                'o_lose_reward': args.o_lose_reward,
             }
         )
         .multi_agent(
@@ -93,18 +103,14 @@ def main():
         )
         .rl_module(
             rl_module_spec=MultiRLModuleSpec(
-                module_specs={
+                rl_module_specs={
                     'pX': RLModuleSpec(
                         module_class=ActionMaskingTorchRLModule,
-                        model_config_dict={
-                            'fcnet_activation': 'relu'
-                        }
+                        model_config=DefaultModelConfig(fcnet_activation='relu'),
                     ),
                     'pO': RLModuleSpec(
                         module_class=ActionMaskingTorchRLModule,
-                        model_config_dict={
-                            'fcnet_activation': 'relu'
-                        }
+                        model_config=DefaultModelConfig(fcnet_activation='relu'),
                     )
                 }
             )
@@ -134,7 +140,7 @@ def main():
             'training_iteration': 'Iters',
         },
         max_report_frequency=10,
-        metric='Tie',
+        metric='Tie', # The metric to sort tables by
         mode='max',
         time_col='Seconds',
         rounding={
@@ -156,7 +162,7 @@ def main():
 
     # Create Tuner Config
     config_tuner = tune.TuneConfig(
-        metric='env_runners/Tie',
+        metric='env_runners/Tie', # Metric to optimize. Pick settings that yield lots of ties.
         mode='max',
         trial_dirname_creator=trail_dirname_creator,
         search_alg=ConcurrencyLimiter(
@@ -165,11 +171,11 @@ def main():
         ),
         scheduler=ASHAScheduler(
             time_attr="training_iterations",
-            grace_period=grace_period_iter,
-            max_t=max_iter_individual,
+            grace_period=grace_period_iter, # Don't stop trials before this point
+            max_t=max_iter_individual,      # Stop trials after this point, no matter what
         ),
         num_samples=-1,
-        time_budget_s=max_time_total,
+        time_budget_s=max_time_total,       # Timeout for all trials
     )
 
     # Create Tuner Object
@@ -180,7 +186,7 @@ def main():
         run_config=train.RunConfig(
             name=experiment_name,
             stop={
-                'env_runners/Tie': 0.99,
+                'env_runners/Tie': args.stop_tie_threshold, # No non-ties within the window
             },
             storage_path=str(PROJECT_PATH / 'results'),
             checkpoint_config=config_checkpoint,
@@ -195,4 +201,13 @@ def main():
 
 
 if __name__ == "__main__":
-    main()
+    # python training.py --x-lose-reward -.1 --o-win-reward .1 --stop-tie-threshold 1.0
+    # Could try asymmetric tie penalties if this doesn't work.
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--x-win-reward", type=float, default=1)
+    parser.add_argument("--x-lose-reward", type=float, default=-1)
+    parser.add_argument("--o-win-reward", type=float, default=1)
+    parser.add_argument("--o-lose-reward", type=float, default=-1)
+    parser.add_argument("--stop-tie-threshold", type=float, default=0.99)
+    args = parser.parse_args()
+    main(args)
